More flexible `sed` strings
Replace `sed` patterns to enhance flexibility, so that the script works for QRs that present the attributes in different order than what the previous patterns assumed. Also, replace `%20` (the URL-encoded representation of a space) with underscores (spaces would be much harder to deal with) to improve the readability of the code list.